% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Provide useful summary statistic of selected variables}
\usage{
check_missing(data, id_time = NULL)
}
\arguments{
\item{data}{data.frame to perform summary statistics.}

\item{id_time}{(Default = \code{NULL}) A column name indicating time (e.g., year) in \code{data}. If unspecified, missing values will counted for the entire data set.}
}
\value{
\code{check_missing} returns a list containing the following elements:
 \itemize{
   \item \code{desc}: Data frame containing descriptive statistics including mean, standard deviation, minimum, maximum, 10, 25, 50, 75, and 90th percentiles for all variables.
   \item \code{miss}: Data frame with total number of missing observations (i.e., countries) for each values in \code{id_time}.
 }
}
\description{
Returns simple descriptive statistics and number of missing entries in each year in the data set.
Character and factor variables are automatically binarized to obtain descriptive statistics.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
