% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stat.R
\name{desc_stat}
\alias{desc_stat}
\title{Provide useful summary statistic of selected variables}
\usage{
desc_stat(data)
}
\arguments{
\item{data}{data.frame to perform summary statistics.}
}
\value{
\code{desc_stat} returns a list containing the following elements:
 \itemize{
   \item \code{desc}: Data frame containing descriptive statistics including mean, standard deviation, minimum, maximum, 10, 25, 50, 75, and 90th percentiles for all variables.
   \item \code{miss}: Data frame with total number of missing observations (i.e., countries) for each year.
   \item \code{corr}: Correlation matrix of all variables.
 }
}
\description{
Returns simple descriptive statistics and number of missing entries in each year in the data set.
Character and factor variables are automatically binarized to obtain descriptive statistics.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
