% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_var.R
\name{impute_var}
\alias{impute_var}
\title{Impute Missing Values}
\usage{
impute_var(
  data,
  id_unit = NULL,
  id_time = NULL,
  var_impute = NULL,
  method = "amelia",
  n_impute = 5,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{id_unit}{A site-level variable name included in \code{data}.}

\item{id_time}{(Default = \code{NULL}). A year variable name included in \code{data}. If not specified, assumes the dataset is cross-sectional.}

\item{var_impute}{(Default = \code{NULL}) Vector with one or more variable names for which imputation is performed. Imputes all variables that contain missing values in \code{data} except \code{id_unit} and \code{id_time} if not specified.}

\item{method}{(Optional. Default = "miceranger") Imputation method. Choose from "amelia", and "miceranger".}

\item{n_impute}{(Optional. Default = 5) The number of imputed datasets to create (equivalent of argument \code{m} in \code{amelia()} and \code{miceRanger()}).}

\item{...}{Arguments passed onto \code{Amelia::amelia()} or \code{miceRanger::miceRanger()}.}
}
\value{
A dataframe with new variables merged onto the base dataset.
}
\description{
Imputes missing values in both numeric and character variables.
Character variables are converted into dummies for each unique values prior to imputation.
The dataset can be either cross-sectional or time-series cross-sectional.
Users may use this function to impute missing values on the site-level variables that they want to diversify prior to subsetting the data to the target population.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
