% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge external dataset}
\usage{
merge_data(
  basedata,
  newdata,
  id_year = NULL,
  id_country = NULL,
  iso3 = FALSE,
  ...
)
}
\arguments{
\item{basedata}{data.frame for new variables to be merged onto. We recommend you use \code{sps_data}.}

\item{newdata}{data.frame that contains new variables. All variables included in the dataset except for the variables asserted in \code{id_country} and \code{id_year} will be merged on to \code{basedata} automatically.}

\item{id_year}{(Default = \code{NULL}) A year variable name included in \code{newdata}.}

\item{id_country}{(Default = \code{NULL}) A country variable name included in \code{newdata}.}

\item{iso3}{(Default = \code{FALSE}) Whether or not the \code{id_country} is an ISO3 country code. If \code{TRUE}, \code{newdata} is merged only using the iso3 code. If \code{FALSE}, then \code{newdata} is merged using exact match, followed by a fuzzy match using \code{fuzzyjoin::stringdist_join()}.}

\item{...}{Arguments passed onto \code{fuzzymatch::stringdist_join()}.}
}
\value{
A dataframe with new variables merged.
}
\description{
Merge external dataset
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
