% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sps_data.R
\name{merge_sps_data}
\alias{merge_sps_data}
\title{Merge new variables from V-Dem/World Bank or external dataset}
\usage{
merge_sps_data(data, vars = NULL)
}
\arguments{
\item{data}{data.frame for new variables to be merged onto.}

\item{vars}{(Default = \code{NULL}) A vector with one or more variable names that should be merged onto \code{data}. Must be specified if \code{newdata} is \code{NULL}. Use \code{search_var()} to search for new variables.}
}
\value{
A dataframe with new variables merged.
}
\description{
Allows users to either add new variable(s) from V-Dem or World Bank or merge their own data set to the baseline data set.
To add variables from V-Dem or World Bank, \code{data}, \code{newdata} and \code{vars} must be supplied.
To merge external data set, \code{data}, \code{newdata}, and \code{id_site} must be supplied.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}. \cr
Michael Coppedge et al. (2023). “V-Dem Dataset V13.” Varieties of Democracy (V-Dem) Project. \url{https://doi.org/10.23696/VDEMDS23}. \cr
World Bank. (2023). World Development Indicators. \url{https://databank.worldbank.org/source/world-development-indicators}.
}
