% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_country_data.R
\name{sps_country_data}
\alias{sps_country_data}
\title{Call particular years from the full country-level dataset}
\usage{
sps_country_data(year = NULL)
}
\arguments{
\item{year}{(Default = \code{NULL}) A year in which the country-level dataset should be subsetted to. Can be either a single numeric value (e.g., 2010) or a range (e.g., 2010:2015). Returns the entire data if \code{NULL}.}
}
\value{
A data.frame with new variables merged onto \code{data}.
}
\description{
Call particular years from the full country-level dataset
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
