% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps_map.R
\name{sps_map}
\alias{sps_map}
\title{Create a world map with target population and selected countries from SPS indicated.}
\usage{
sps_map(data, targetvar = NULL, countryvar = NULL, countrysps = NULL)
}
\arguments{
\item{data}{Data frame where \code{targetvar} and \code{countryvar} are stored.}

\item{targetvar}{Name of the binary indicator variable in \code{data} that indicates target population. Default is NULL, which treats all countries as target population.}

\item{countryvar}{Name of the country variable in \code{data}.}

\item{countrysps}{A character vector including country names selected from SPS. Values must come from the variable selected in \code{countryvar}.}
}
\value{
A dataframe with new variables merged.
}
\description{
Create a world map with target population and selected countries from SPS indicated.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
