% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{add_var}
\alias{add_var}
\title{Add Additional Covariates from V-Dem Dataset and/or World Bank API}
\usage{
add_var(data, vars = NULL)
}
\arguments{
\item{data}{Dataframe for new variables to be merged onto (we recommend you use \code{sps_data}). A dataset must have a column 'iso3' and 'year'.}

\item{vars}{A character vector with one or more variable names from V-Dem dataset or World Bank Indicators.}
}
\value{
A dataframe with new variables merged on
}
\description{
Add Additional Covariates from V-Dem Dataset and/or World Bank API
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
