% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var.R
\name{add_var}
\alias{add_var}
\title{Synthetic Purposive Sampling: Dataset for Site Selection
Add Additional Covariates from V-Dem Dataset and/or World Bank API}
\usage{
add_var(data = sps_data, vars = NULL)
}
\arguments{
\item{data}{Dataframe for new variables to be merged onto. Default is sps_data. A dataset must have a column 'iso3' and 'year'.}

\item{vars}{A character vector with one or more variable names from V-Dem dataset or World Bank Indicators.}
}
\value{
A dataframe with new variables merged on
}
\description{
Synthetic Purposive Sampling: Dataset for Site Selection
Add Additional Covariates from V-Dem Dataset and/or World Bank API
}
\examples{
data(sps_data)
new <- add_var(vars = "SM.POP.TOTL.ZS")
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
