% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_var.R
\name{impute_var}
\alias{impute_var}
\title{Impute Missing Values}
\usage{
impute_var(
  data,
  countryvar = NULL,
  yearvar = NULL,
  imputevars = NULL,
  method = "carryforward",
  carryforward_yrs = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe for new variables to be merged onto (we recommend you use \code{sps_data}).}

\item{countryvar}{Country column name in \code{data}. Default is "iso3".}

\item{yearvar}{Year column name in \code{data}. Default is "year".}

\item{imputevars}{A character vector with one or more variable names for which imputation is performed.}

\item{method}{Imputation method. Choose from "carryforward" (default), "amelia", and "miceranger".}

\item{carryforward_yrs}{Maximum number of years allowed to be carried forward. Used only when "carryforward" is selected for \code{method}. Default is NULL, which will impute values from the most recent years available.}

\item{...}{Arguments passed onto \code{Amelia::amelia} or \code{miceRanger::miceRanger}.}
}
\value{
A dataframe with new variables merged on
}
\description{
Impute Missing Values
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
