% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_var.R
\name{impute_var}
\alias{impute_var}
\title{Impute Missing Values}
\usage{
impute_var(
  data,
  id_site = NULL,
  id_year = NULL,
  id_impute = NULL,
  method = "miceranger",
  n_impute = 5,
  carryforward_yrs = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{id_site}{A site-level variable name included in \code{data}.}

\item{id_year}{A year variable name included in \code{data}.}

\item{id_impute}{(Default = \code{NULL}) Vector with one or more variable names for which imputation is performed. Imputes all variables in \code{data} except \code{id_site} and \code{id_year} if not specified.}

\item{method}{(Optional. Default = "miceranger") Imputation method. Choose from "carryforward", "amelia", and "miceranger".}

\item{n_impute}{(Optional. Used only when "miceranger" or "amelia" is selected in \code{method}. Default = 5) The number of imputed datasets to create (equivalent of argument \code{m} in \code{amelia()} and \code{miceRanger()}).}

\item{carryforward_yrs}{(Optional. Used only when "carryforward" is selected in \code{method}. Default = \code{NULL}) Maximum number of years allowed to be carried forward. If not specified, it will impute values from the most recent years available.}

\item{...}{Arguments passed onto \code{Amelia::amelia()} or \code{miceRanger::miceRanger()}.}
}
\value{
A dataframe with new variables merged onto the base dataset.
}
\description{
Impute Missing Values
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
