% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_var.R
\name{impute_var}
\alias{impute_var}
\title{Impute Missing Values}
\usage{
impute_var(
  data,
  id_unit = NULL,
  id_time = NULL,
  var_impute = NULL,
  var_predictor = NULL,
  var_ord = NULL,
  var_nom = NULL,
  var_lgstc = NULL,
  method = "amelia",
  n_impute = 5,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing variables to impute.}

\item{id_unit}{A unique identifier for sites. A column name in \code{data}.}

\item{id_time}{(Default = \code{NULL}) A unique identifier for time index. A column name in \code{data}. If unspecified, it assumes \code{data} is cross-sectional.}

\item{var_impute}{(Default = \code{NULL}) A vector with one or more variable names for which imputation is performed. Imputes all variables that contain missing values in \code{data} except \code{id_unit} and \code{id_time} if not specified.}

\item{var_predictor}{(Default = \code{NULL}) A vector with one or more variable names that we use as predictors to impute variables in \code{var_impute}. If \code{NULL}, the function uses all variables in \code{data} except for variables in \code{var_impute}.}

\item{var_ord}{(Default = \code{NULL}) A vector with one or more variable names to be imputed which are ordinal. Binary variable names can be in either \code{var_ord} or \code{var_nom}.}

\item{var_nom}{(Default = \code{NULL}) A vector with one or more variable names to be imputed which are nominal.}

\item{var_lgstc}{(Default = \code{NULL}) A vector with one or more variable names to be imputed which are proportional (ranging between 0 and 1).}

\item{method}{(Optional. Default = "amelia") Imputation method. Choose from "amelia", "mice", and "miceranger".}

\item{n_impute}{(Optional. Default = 5) The number of imputed datasets to create (equivalent of argument \code{m} in \code{amelia()}, \code{mice()}, and \code{miceRanger()}).}

\item{...}{Arguments passed onto \code{Amelia::amelia()}, \code{mice::mice()}, or \code{miceRanger::miceRanger()}.}
}
\value{
A dataframe with new variables merged onto the base dataset.
}
\description{
Imputes missing values in both numeric and character variables.
Character variables are converted into dummies for each unique values prior to imputation.
The dataset can be either cross-sectional or time-series cross-sectional.
Users may use this function to impute missing values on the site-level variables that they want to diversify prior to subsetting the data to the target population.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
