% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge external dataset}
\usage{
merge_data(basedata, newdata, id_site = NULL, id_year = NULL, ...)
}
\arguments{
\item{basedata}{data.frame for new variables to be merged onto.}

\item{newdata}{data.frame that contains new variables. All variables included in the dataset except for the variables asserted in \code{id_site} and \code{id_year} will be merged on to \code{basedata} automatically.}

\item{id_site}{(Default = \code{NULL}) A site variable name included in \code{basedata} and \code{newdata}.}

\item{id_year}{(Optional, Default = \code{NULL}) A year variable name included in \code{basedata} and \code{newdata}. If \code{NULL}, assumes \code{newdata} is cross-sectional.}

\item{...}{Arguments passed onto \code{fuzzymatch::stringdist_join()}.}
}
\value{
A dataframe with new variables merged.
}
\description{
Merge external dataset
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
