% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge external dataset}
\usage{
merge_data(
  basedata,
  newdata = NULL,
  yearvar = NULL,
  countryvar = NULL,
  iso = FALSE
)
}
\arguments{
\item{basedata}{Data frame for new variables to be merged onto. We recommend you use \code{sps_data}.}

\item{newdata}{Data frame that contains new variables. All variables included in the dataset except for the variables asserted in \code{countryvar} and \code{yearvar} will be merged on to \code{basedata} automatically.}

\item{yearvar}{Year column name in \code{newdata}.}

\item{countryvar}{Country column name in \code{newdata}.}

\item{iso}{Whether or not the \code{countryvar} is an ISO3 code. If TRUE, the \code{newdata} is merged only using the iso3 code. If FALSE, the \code{newdata} is merged using exact match, followed by a probabilistic match using \code{fastLink} Default is FALSE.}

\item{fuzz_match}{Type of fuzzy match to be conducted. Choose from "fastLink" or "fuzzyjoin".}
}
\value{
A dataframe with new variables merged.
}
\description{
Merge external dataset
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}.
}
