% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_var.R
\name{search_var}
\alias{search_var}
\title{Search for Additional Covariates from V-Dem Dataset and World Bank API}
\usage{
search_var(keyword = NULL)
}
\arguments{
\item{keyword}{A character string with a keyword to search for. Also allows for regex notation.}
}
\value{
A dataframe with list of variable names, descriptions, variable source, and coverage.
}
\description{
Allows users to search for other country-level variables in V-Dem and World Bank using keywords.
}
\references{
Egami and Lee. (2023+). Designing Multi-Context Studies for External Validity: Site Selection via Synthetic Purposive Sampling. Available at \url{https://naokiegami.com/paper/sps.pdf}. \cr
Michael Coppedge et al. (2023). “V-Dem Dataset V13.” Varieties of Democracy (V-Dem) Project. \url{https://doi.org/10.23696/VDEMDS23}.\cr
World Bank. (2023). World Development Indicators. \url{https://databank.worldbank.org/source/world-development-indicators}.
}
