% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sps_country_data}
\alias{sps_country_data}
\title{SPS country-level site selection data}
\format{
data
}
\description{
A country-level data set comprised of 21 variables for 179 countries between 2010 and 2022.
It contains a wide range of variables commonly used in the social sciences: they represent political, economic, and social characteristics of countries.
}
\keyword{datasets}
